extract disruption builder and pass context
